Create 26. 从排序数组中删除重复项.c
Given an integer array sorted in non-decreasing order, remove the duplicates in-place such that each unique element appears only once. The relative order of the elements should be kept the same.nums

Since it is impossible to change the length of the array in some languages, you must instead have the result be placed in the first part of the array . More formally, if there are elements after removing the duplicates, then the first elements of  should hold the final result. It does not matter what you leave beyond the first  elements.numskknumsk

Return after placing the final result in the first slots of .kknums

Do not allocate extra space for another array. You must do this by modifying the input array in-place with O(1) extra memory.